# openapi.yaml
openapi: 3.0.2
info:
    title: Sunflower Log
    version: 1.0.0
    description: A service to log all activities on Sunflower Labs apps.
    contact:
        name: Nuno Barreto
        email: nuno@sunflower-labs.com
servers:
    -
        url: /api/v1
        description: ''
paths:
    /authenticate:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AuthenticationClient'
                required: true
            responses:
                '200':
                    description: Authentication token correctly received.
                '400':
                    description: The body parameters client_id and client_secret are mandatory.
                '401':
                    description: Request failed with status code 401.
    /log:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LogEntry'
                required: true
            responses:
                '201':
                    description: New log entry created.
components:
    schemas:
        LogEntry:
            title: Root Type for LogEntry
            description: ''
            required:
                - message
            type: object
            properties:
                level:
                    description: ''
                    enum:
                        - emerg
                        - alert
                        - crit
                        - error
                        - warning
                        - notice
                        - info
                        - debug
                    type: string
                message:
                    properties:
                        anything:
                            type: string
            example:
                level: error
                message:
                    anything: is possible
        AuthenticationClient:
            title: Root Type for AuthenticationClient
            description: ''
            type: object
            properties:
                client_id:
                    type: string
                client_secret:
                    type: string
            example:
                client_id: SOME_CLIENT_ID
                client_secret: SOME_CLIENT_SECRET